+++ lib/src/core/unique_string_generator.dart
@@ -1,0 +1,9 @@
+import 'dart:math';
+
+String getUniqueString() {
+  final random = Random();
+  final randomNumber = random.nextInt(10000000).toString();
+  final timeStamp = DateTime.now().toIso8601String();
+
+  return "$randomNumber$timeStamp";
+}