+++ lib/src/widgets/toast_slider.dart
@@ -6,13 +6,17 @@ import 'toast_position_widget.dart';
 import 'toast_progress_bar_widget.dart';
 
 class ToastSlider extends StatefulWidget {
-  /// The overlay entry to show the toast
-  final OverlayEntry overlayEntry;
+  /// The toast controller for removing the overlay
+  final ToastController toastController;
 
-  /// The title is a widget displayed at the left side of the toast
+  /// The widget to display at the start of the message
+  final Widget? leading;
+
+  /// The Message to be displayed at the toast
+  /// It's width is expanded whenever leading or trailing widget is present
   final Widget title;
 
-  /// The trailing is a widget displayed at the right side of the toast
+  /// A widget displayed at the right side of the toast
   final Widget? trailing;
 
   /// The setting for the sliding animation
@@ -27,20 +31,17 @@ class ToastSlider extends StatefulWidget {
   /// Function to be called when the toast is clicked
   final Function()? onTapped;
 
-  /// Call callback only one time
-  final bool disableMultiTapping;
-
   /// A sliding message to show the toast
   const ToastSlider({
     super.key,
-    required this.overlayEntry,
+    required this.toastController,
+    required this.leading,
     required this.title,
     required this.trailing,
     required this.toastSetting,
     required this.toastStyle,
     this.onDisposed,
     this.onTapped,
-    required this.disableMultiTapping,
   });
 
   @override
@@ -65,11 +66,13 @@ class _ToastSliderState extends State<ToastSlider>
     toastSetting = widget.toastSetting;
     toastStyle = widget.toastStyle;
 
+    /// For sliding effect
     slideController = AnimationController(
       vsync: this,
       duration: toastSetting.animationDuration,
     );
 
+    /// For displaying the toast and also for progress bar
     sizeController = AnimationController(
       vsync: this,
       duration: toastSetting.animationDuration + toastSetting.displayDuration,
@@ -107,7 +110,7 @@ class _ToastSliderState extends State<ToastSlider>
       // Wait for the reverse animation to complete
       if (toastSetting.showReverseAnimation) await slideController.reverse();
       // Remove the overlay entry
-      widget.overlayEntry.remove();
+      widget.toastController.closeToast();
       // Execute the on disposed function
       widget.onDisposed?.call();
     }
@@ -124,6 +127,7 @@ class _ToastSliderState extends State<ToastSlider>
   Widget build(BuildContext context) {
     // Create the toast container
     Widget child = ToastContainerWidget(
+      leading: widget.leading,
       title: widget.title,
       trailing: widget.trailing,
       toastStyle: toastStyle,
@@ -171,13 +175,13 @@ class _ToastSliderState extends State<ToastSlider>
         // Stop animations and remove the overlay on dismissed
         onDismissed: (_) {
           sizeController.stop();
-          widget.overlayEntry.remove();
+          widget.toastController.closeToast();
         },
         child: GestureDetector(
           // Execute onTapped function on tap
           onTap: () {
             if (!isToastTapped) widget.onTapped?.call();
-            if (widget.disableMultiTapping) isToastTapped = true;
+            if (widget.toastSetting.disableMultiTapping) isToastTapped = true;
           },
           // Pause the animation on long press
           onLongPress: () => sizeController.stop(),